functiondisplay_validators_info moved to common
